First version

+++ Management/Clients/Models/AirQualityCityResponse.cs
@@ -55,19 +55,19 @@ namespace Management.Clients.Models
         public string Ts { get; set; }
 
         [JsonProperty("tp")]
-        public int Tp { get; set; }
+        public double Tp { get; set; }
 
         [JsonProperty("pr")]
-        public int Pr { get; set; }
+        public double Pr { get; set; }
 
         [JsonProperty("hu")]
-        public int Hu { get; set; }
+        public double Hu { get; set; }
 
         [JsonProperty("ws")]
-        public int Ws { get; set; }
+        public double Ws { get; set; }
 
         [JsonProperty("wd")]
-        public int Wd { get; set; }
+        public double Wd { get; set; }
 
         [JsonProperty("ic")]
         public string Ic { get; set; }
@@ -79,13 +79,13 @@ namespace Management.Clients.Models
         public string Ts { get; set; }
 
         [JsonProperty("aqius")]
-        public int Aqius { get; set; }
+        public double Aqius { get; set; }
 
         [JsonProperty("mainus")]
         public string Mainus { get; set; }
 
         [JsonProperty("aqicn")]
-        public int Aqicn { get; set; }
+        public double Aqicn { get; set; }
 
         [JsonProperty("maincn")]
         public string Maincn { get; set; }
